Use consistent location for icons
