+++ frontend/src/components/dashboard/Tips.tsx
@@ -4,8 +4,7 @@ import Link from "next/link";
 import { useTabList, useTabPanel, useTab } from "react-aria";
 import { useTabListState, TabListState, Item } from "react-stately";
 import styles from "./Tips.module.scss";
-import { CloseIcon } from "../icons/close";
-import { InfoIcon } from "../icons/info";
+import { InfoIcon, CloseIcon } from "../Icons";
 import { ProfileData } from "../../hooks/api/profile";
 import { useLocalDismissal } from "../../hooks/localDismissal";
 import { getRuntimeConfig } from "../../config";
