+++ frontend/src/components/layout/NpsSurvey.tsx
@@ -5,7 +5,7 @@ import { useLocalDismissal } from "../../hooks/localDismissal";
 import { useIsLoggedIn } from "../../hooks/session";
 import { useLocalization } from "@fluent/react";
 import { useProfiles } from "../../hooks/api/profile";
-import { CloseIcon } from "../icons/close";
+import { CloseIcon } from "../Icons";
 
 /**
  * Quickly survey the user for input to our Net Promotor Score.
